+++ lua/plugins/git.lua
@@ -74,19 +74,4 @@ return {
     },
   },
 
-  -- Neogit disabled - too slow on Windows (~3s to open)
-  -- {
-  --   "NeogitOrg/neogit",
-  --   lazy = true,
-  --   dependencies = {
-  --     "nvim-lua/plenary.nvim",
-  --     "sindrets/diffview.nvim",
-  --     "nvim-telescope/telescope.nvim",
-  --     "folke/snacks.nvim",
-  --   },
-  --   cmd = "Neogit",
-  --   keys = {
-  --     { "<leader>gn", "<cmd>Neogit<cr>", desc = "Neogit UI" }
-  --   }
-  -- },
 }